Create new branch & Use Django Rest Framework & update requirements.txt
I've created new branch for the new features testing.
DRF is now in the project requirements.
The reason is that we need to implement front-to-back api and cross-platform api in further development of the project.
So, it is better to start using it now from the very beginning.

Implemented debug user API.
Created "/users/" and "/users/<int:id>" endpoints (only for the debug purposes and showing the capabilities of the DRF).

Included necessary base libraries to the requirements.txt that are helpful for the further development.